correct a compiler's warning...

+++ src/lib/RVNGZipStream.cpp
@@ -489,7 +489,7 @@ RVNGInputStream *RVNGZipStream::getSubstream(RVNGInputStream *input, const char 
 
 		if (strm.total_out == 0)
 			return 0;
-		return new RVNGStringStream(&data[0], strm.total_out);
+		return new RVNGStringStream(&data[0], (unsigned int) strm.total_out);
 	}
 }
